Add exports for inspector and shared ui components needed for extending inspector v2
