+++ packages/dev/inspector-v2/src/components/scene/sceneExplorer.tsx
@@ -560,7 +560,12 @@ export const SceneExplorer: FunctionComponent<{
         visibleItems.add(sceneTreeItem);
 
         for (const sectionTreeItem of sectionTreeItems) {
+            const children = sectionTreeItem.children;
             traversedItems.push(sectionTreeItem);
+            if (!children.length) {
+                continue;
+            }
+
             // Section tree items are always visible when not filtering.
             if (!filter) {
                 visibleItems.add(sectionTreeItem);
@@ -569,7 +574,7 @@ export const SceneExplorer: FunctionComponent<{
             // When an item filter is present, always traverse the full scene graph (e.g. ignore the open item state).
             if (filter || openItems.has(sectionTreeItem.sectionName)) {
                 TraverseGraph(
-                    sectionTreeItem.children,
+                    children,
                     // Get children
                     (treeItem) => {
                         if (filter || openItems.has(treeItem.entity.uniqueId)) {
