+++ packages/dev/inspector-v2/src/index.ts
@@ -1,1 +1,61 @@
+// Export the parts of inspector that are intended to be part of the public API.
+export * from "./components/properties/boundProperty";
+export * from "./components/properties/linkToEntityPropertyLine";
+export { EntityBase, EntityDisplayInfo, SceneExplorerCommand, SceneExplorerCommandProvider, SceneExplorerSection } from "./components/scene/sceneExplorer";
+export * from "./components/extensibleAccordion";
+export { Pane as PaneContainer } from "./components/pane";
+export * from "./components/teachingMoment";
+export * from "./extensibility/extensionFeed";
+export * from "./hooks/compoundPropertyHooks";
+export * from "./hooks/instrumentationHooks";
+export * from "./hooks/observableHooks";
+export * from "./hooks/pollingHooks";
+export * from "./hooks/resourceHooks";
+export * from "./hooks/settingsHooks";
+export * from "./hooks/teachingMomentHooks";
+export * from "./instrumentation/functionInstrumentation";
+export * from "./instrumentation/propertyInstrumentation";
+export * from "./modularity/serviceDefinition";
+export { IPropertiesService, PropertiesServiceIdentity } from "./services/panes/properties/propertiesService";
+export { ISceneExplorerService, SceneExplorerServiceIdentity } from "./services/panes/scene/sceneExplorerService";
+export { IDebugService, DebugServiceIdentity } from "./services/panes/debugService";
+export { ISettingsService, SettingsServiceIdentity } from "./services/panes/settingsService";
+export { IStatsService, StatsServiceIdentity } from "./services/panes/statsService";
+export { IToolsService, ToolsServiceIdentity } from "./services/panes/toolsService";
+export * from "./services/sceneContext";
+export * from "./services/selectionService";
+export * from "./services/settingsContext";
+export { IShellService, ToolbarItem, SidePane, CentralContent, ShellServiceIdentity } from "./services/shellService";
 export * from "./inspector";
+
+// Export the shared primitive UI controls that can be used for extending the inspector.
+export * from "shared-ui-components/fluent/primitives/accordion";
+export * from "shared-ui-components/fluent/primitives/button";
+export * from "shared-ui-components/fluent/primitives/checkbox";
+export * from "shared-ui-components/fluent/primitives/collapse";
+export * from "shared-ui-components/fluent/primitives/colorPicker";
+export * from "shared-ui-components/fluent/primitives/comboBox";
+export * from "shared-ui-components/fluent/primitives/draggable";
+export * from "shared-ui-components/fluent/primitives/dropdown";
+export * from "shared-ui-components/fluent/primitives/gradient";
+export * from "shared-ui-components/fluent/primitives/infoLabel";
+export * from "shared-ui-components/fluent/primitives/lazyComponent";
+export * from "shared-ui-components/fluent/primitives/link";
+export * from "shared-ui-components/fluent/primitives/list";
+export * from "shared-ui-components/fluent/primitives/messageBar";
+export * from "shared-ui-components/fluent/primitives/positionedPopover";
+export * from "shared-ui-components/fluent/primitives/primitive";
+export * from "shared-ui-components/fluent/primitives/searchBar";
+export * from "shared-ui-components/fluent/primitives/searchBox";
+export * from "shared-ui-components/fluent/primitives/spinButton";
+export * from "shared-ui-components/fluent/primitives/switch";
+export * from "shared-ui-components/fluent/primitives/syncedSlider";
+export * from "shared-ui-components/fluent/primitives/textarea";
+export * from "shared-ui-components/fluent/primitives/textInput";
+export * from "shared-ui-components/fluent/primitives/toggleButton";
+
+// Export the shared hoc UI controls that can be used for extending the inspector.
+export * from "shared-ui-components/fluent/hoc/buttonLine";
+export * from "shared-ui-components/fluent/hoc/fileUploadLine";
+export * from "shared-ui-components/fluent/hoc/gradientList";
+export * from "shared-ui-components/fluent/hoc/pane";
