+++ packages/dev/sharedUiComponents/src/fluent/primitives/utils.ts
@@ -1,0 +1,15 @@
+import type { KeyboardEvent, FocusEvent } from "react";
+
+export function HandleOnBlur(event: FocusEvent<HTMLInputElement>) {
+    event.stopPropagation();
+    event.preventDefault();
+}
+
+export function HandleKeyDown(event: KeyboardEvent<HTMLInputElement>) {
+    event.stopPropagation(); // Prevent event propagation
+
+    // Prevent Enter key from causing form submission or value reversion
+    if (event.key === "Enter") {
+        event.preventDefault();
+    }
+}
